Used appropriate data types for Instruction members

+++ cpu/instruction.cpp
@@ -4,7 +4,7 @@ bool Instruction::isExtendedInstruction(uint8_t firstOpcodeByte) {
   return firstOpcodeByte == 0xCB;
 }
 
-Instruction::Instruction(unsigned int ticks, unsigned int dataSize, unsigned int opcodeSize):
+Instruction::Instruction(unsigned long long ticks, unsigned short dataSize, unsigned short opcodeSize):
   ticks(ticks),
   dataSize(dataSize),
   opcodeSize(opcodeSize)
